app/requirements.txt
pandas==2.2.1
pyyaml==6.0.1
pyarrow==15.0.2
argparse==1.4.0